Upgrade zephyr to 3.7.0
* Disable support for ipv6
* Rework net interfaces initialization
* Increase number of available eventfd
* Improve retry mecanism to acquire net time
* Fix compilation for QEMU
* Remove deprecated features, fix warnings

Remaining tasks:
* Network does not work with QEMU (especially sntp). This needs to be investigated.

+++ src/http_server/prometheus_client.c
@@ -232,17 +232,17 @@ static ssize_t encode_value(char *buf, size_t buf_size, struct metric_value *val
 		ret = snprintf(buf, buf_size, "%u", value->uvalue);
 		break;
 	case VALUE_ENCODING_TYPE_FLOAT_DIGITS:
-		ret = snprintf(buf, buf_size, "%.*f", (int)value->encoding.digits, value->fvalue);
+		ret = snprintf(buf, buf_size, "%.*f", (int)value->encoding.digits, (double)value->fvalue);
 		break;
 	case VALUE_ENCODING_TYPE_EXP:
-		ret = snprintf(buf, buf_size, "%e", value->fvalue);
+		ret = snprintf(buf, buf_size, "%e", (double)value->fvalue);
 		break;
 	case VALUE_ENCODING_TYPE_EXP_DIGITS:
-		ret = snprintf(buf, buf_size, "%.*e", (int)value->encoding.digits, value->fvalue);
+		ret = snprintf(buf, buf_size, "%.*e", (int)value->encoding.digits, (double)value->fvalue);
 		break;
 	case VALUE_ENCODING_TYPE_FLOAT:
 	default:
-		ret = snprintf(buf, buf_size, "%f", value->fvalue);
+		ret = snprintf(buf, buf_size, "%f", (double)value->fvalue);
 		break;
 	}
